Allow configuring the max number of redis connections
The sidekiq processor runs as many threads as there are CPU cores
available. If this number exceeds the maximum number of redis
connections available on the pool, additional connection requests, e.g.
in the `_health` endpoint, may timeout.

Set the max redis connections to be the cpu count, plus a small buffer.