Add CLI commands to run DB migrations
This prompted a refactor of the `RunCommand` trait to add `A: App` and
pass the `App` instance as a parameter in `RunCommand::run`. This is
possibly useful for other commands, and potentially for the consumer's
custom commands as well. Another benefit of passing the `App` instance
as a parameter is it simplifies the syntax for specifying the `A`
generic argument.

https://github.com/roadster-rs/roadster/issues/41