Remove stale periodic jobs
Sidekiq.rs only provides the ability to clear all periodic jobs. If this
method is used to remove stale periodic jobs when deploying, then the
periodic jobs are repeatedly removed and re-added, even when they aren't
stale.

Add a way to track which periodic workers are added, and once all
periodic jobs are added, remove any that were added by a previous
deployment but not the current deployment. This way, only the actually
stale periodic jobs are removed.