Use a separate redis connection pool for enqueuing vs fetching jobs
Problem
-------
If the job queue is empty, sidekiq.rs will wait a couple seconds before
releasing its connection. This can delay adding items to the queue by
up to a couple seconds depending on how the timing works out, which is
bad because adding the the queue should be as fast as possible.

Solution
--------
Use separate connection pools for enqueuing vs fetching jobs from the
queue. This decouples enqueuing from needing to wait for a connection
that is being blocked by a worker task that's waiting for a job to be
added to the queue in the `sidekiq::Processor`.