Add `AppService` trait and use it to implement the HTTP service
Add an `AppService` trait to allow consumers to define custom services
that will be run as tokio tasks. This is similar (I believe) to
Laraval's `Provider` concept.

Use the `AppService` trait to add an HTTP server using Axum instead of
putting http/axum specific methods directly on the `App` trait

- This will help with https://github.com/roadster-rs/roadster/issues/9
  (and we may be able to resolve it)
- This will resolve https://github.com/roadster-rs/roadster/issues/19