Add methods to AppContext instead of direct field access
This makes it slightly more ergonomic to contain the app context state
with an "inner" type that's wrapped by an Arc.

This also allows consumers to disallow direct access to specific fields
using clippy's
[disallowed-methods](https://rust-lang.github.io/rust-clippy/master/index.html#/disallowed_methods)
lint. This is useful, for example, if consumers want to provide their
own method for accessing the DB that performs some logic before/after
the DB operations, e.g. for implementing RLS.