Use automock to mock traits instead of the mock! macro
There were a couple places we were using `mock!` instead of `automock`.
IIRC this was because `automock` wasn't working for `async_trait`
traits; however, after reading the `mockall` docs, I found that
`automock` needs to be placed _before_ `async_trait` in order for it to
work. Also, specifically for the `App` trait, I just learned how to
specify associated types with `automock`.

With this new knowledge, we can replace a couple manual `mock!` impls
with `automock`.