Allow partial overrides of all configs
The app config defines defaults for a lot of things. However,
if the user provides a field, sometimes that means the defaults
for the rest of that sub-config don’t use the default and need
to be set by the user. This is especially difficult to work
around for fields that re-use the same struct.

Refactor how the app config fields that share a common struct are
contructed to provide a different default for each use of the struct,
while still allowing the user to partially override individual fields of
the config.

Also, add tests for all (most?) of the app config to ensure this
behavior is not broken in the future.

Closes https://github.com/roadster-rs/roadster/issues/152