fix: Invalid env var separator on bash
We use `.` as the env var separator in order to differentiate between the prefix
and different levels of the config vs individual config field names
(which would be in UPPER_SNAKE_CASE). However, `.` is not a valid bash
identifier, so we need to use something different.

Switch to using a double underscore (`__`) instead of `.`. This way, we can
support bash and still be able to differentiate between the prefix
and different levels of the config vs individual config field names.

Closes: https://github.com/roadster-rs/roadster/issues/219