feat: Add health checks to run before starting services
This is a separate trait, vs adding a "health check" method to `AppService`, to allow defining
health checks that apply to multiple services. For example, most services would require
the DB and Redis connections to be valid, so we would want to perform a check for these
resources a single time before starting any service instead of once for every service that
needs the resources.

Another benefit of using a separate trait is, because the health checks are decoupled from
services, they can potentially be used in other parts of the app. For example, they could
be used to implement a "health check" API endpoint.

For now, I think it makes sense to register health checks on the
`ServiceRegistry` since the checks are run right before running the
services, and the checks need to succeed in order to run the services.
In the future, it may make more sense to add a new method to the `App`
trait in order to register health checks. In general, however, I'm
trying to steer away from adding too many methods to the `App` trait.

Closes https://github.com/roadster-rs/roadster/issues/237