feat: Add `TestCase` utility for configuring `insta` settings (#277)
Add some utilities to set some common `insta` settings. The settings are
configurable using the `TestCaseConfig` builder struct. The builder
yields a `TestCase` struct containing the configured `insta` settings.

The settings available to configure currently are:

1. Redact UUIDs from snapshots; UUIDs are replaced with the string
`[uuid]`
2. Set the suffix of the snapshot (which can be provided manually or
automatically based on the thread name). This is useful when using
`insta` along with `rstest`.
3. Auto-bind to the current scope on creation of the `TestCase` struct
to avoid the need to remember to bind the settings scope.