feat: Use IDENTITY column for int primary keys instead of BIGSERIAL (#293)
`BIGSERIAL` is not recommended. See:
https://wiki.postgresql.org/wiki/Don%27t_Do_This#Don.27t_use_serial. The
recommendation is instead to use `IDENTITY` columns.

This PR updates the create `user` table migration (just the one using
bigint as the primary key) to create the pk as an IDENTITY column
instead of a BIGSERIAL. This PR also adds the `pk_bigint_identity`
method to enable creating an IDENTITY pk in any table. The
`pk_bigint_identity_options` method is added as well, which is the same
except it allows configuring the sequence for the identity generation.
See: https://www.postgresql.org/docs/current/sql-createsequence.html

Closes https://github.com/roadster-rs/roadster/issues/290