feat: Add middleware to log the request/response payloads (#304)
It can be useful when developing an application to be able to see the
request/response bodies as they are seen by the server. Add an Axum
middleware to do this. Because the middleware buffers the entire
request/response, it is disabled by default. If we add a mechanism to
provide different defaults per environment, this middleware may be
enabled in non-prod environments by default in the future.