feat: Allow jwt from cookie, but only if it's explicitly requested
Getting the JWT from the cookie seems to be acceptable, and is useful
for GET requests (e.g., initial page load when using SSR). However, for
POST requests it should be accompanied with some CSRF protection
mechanism. At the moment, the CSRF validation will need to be performed
by the consuming application.