feat: Remove cookie extraction for `Jwt`, but allow it in `JwtCsrf`
Add a new axum extractor called `JwtCsrf` that allows extracting the JWT
from a cookie (if the `auth.jwt.cookie-name` config is set). This
extractor provides some metadata about whether it's safe to use the JWT
or if a CSRF protection mechanism needs to be applied first.

With this, we also remove cookie extraction from the normal `Jwt`, so it
can safely be used without any CSRF protection mechanisms.