feat!: App methods take `self`
It may be useful for consumers to be able to put state on
their App implementation that they can use in the app trait.
For example, I wanted this when I was working on generating an OpenAPI
client from the app's Aide OpenAPI schema.

Also, this would be needed in order to support a builder-style approach
to building the app. This would allow us to provide a default app impl
configured with a builder-style api. We may still allow consumers to
directly use the trait impl if they want.

Closes https://github.com/roadster-rs/roadster/issues/272