feat: Add more builder methods for `RoadsterApp`
Allow providing most fields of the `RoadsterApp` directly instead of in
provider callback methods. This can be useful if the consumer doesn't
need the app context/state to build an item and they don't want to deal
with creating a callback lambda/method.

The `app-builder` example was updated to use all the new methods.