feat: Enable consumers to provide custom Environment values (#439)
Consumers may want to create custom enviornments, e.g. `gamma` or
`staging`.

Add a `Environment::Custom` enum variant. The custom value is contained
in a String in the variant.

Note: A future release may remove the `From<Environment> for &'static
str` and `From<&Environment> for &'static str` implementations because
it's not possible to convert `Environment::Custom` to a static str. It's
kept for now because it was implemented before we added the
`Environment::Custom` variant.

Closes https://github.com/roadster-rs/roadster/issues/438