feat: Add `AnyMiddleware` to minimize boilerplate for Axum middleware
If a consumer wants to use a middleware that's not already supported by
Roadster, they need to implement the `Middleware` trait, which is extra
boilerplate that could be annoying.

Add `AnyMiddleware` struct that implements the `Middleware` trait, so
consumers just need to provide the name of the middleware and the logic
to build/configure it.

Closes https://github.com/roadster-rs/roadster/issues/470