feat: Add `AnyInitializer` to minimize boilerplate for initializers
If a consumer wants to initialize the Axum Router in a way that's not
already supported by Roadster, they need to implement the `Initializer`
trait, which is extra boilerplate that could be annoying.

Add `AnyInitializer` struct that implements the `Initializer` trait, so
consumers just need to provide the name of the initializer and the logic
to build/configure the Axum Router.

Note: This PR also contains a breaking change to `AnyMiddleware`. However,
since this was just released a few hours ago, I doubt anyone is using it
yet, so it should be relatively safe to just release the breaking change
in a patch update.

Closes https://github.com/roadster-rs/roadster/issues/473