feat!: Enable fetching concrete service from registry via downcast
Add ability to fetch a concrete `AppService` reference from the
`ServiceRegistry`. This enables getting a specific `AppService` in order
to call methods that only exist on that concrete service. For example,
in order to call `HttpService::open_api` on a registered service,

```rust
let service = registry.get::<HttpService>().unwrap();
let api = service.open_api();
```

See `ServiceRegistry::get` for more details and an example.

This PR also adds some OpenAPI related methods to `HttpService`.

Both of these things combined can be used in a build system to generate
the OpenAPI schema for the app and subsequently generate an OpenAPI
client without needing to actually start the service. This is
particularly useful if using Roadster with a non-rust frontend (e.g.
React) or a Rust frontend that does not provide an RPC mechanism (e.g.
Yew). For a Rust frontend that _does_ provide an RPC mechanism (and
therefore does not need this OpenAPI mechanism), see
[Leptos](https://github.com/leptos-rs/leptos).