chore: Refactor `RoadsterApp` to reduce duplication
Use conditional compilation of type parameters of `RoadsterApp` to
enable/disable the bounds on the type parameters depending on which
features are enabled. This allows us to have a single implementation of
`RoadsterApp` instead of having multiple implementations that are
conditionally compiled using `cfg_if`, greatly reducing code duplication
and our maintenance burden.

A downside of this approach is RustRover seems to not play super well
with the conditionally compiled type parameters. It seems to sometimes
ignore the conditional compilation, leading to it thinking that there
are 5 type parameters even though two sets of them are mutually
exclusive and can't be enabled at the same time so there should only
ever be 3 type parameters. However, things compile and we can write our
code in a certain way that RustRover seems to like. See the
`app-builder` example for an example.

Also, add an `Empty` type that can be used as a placeholder/default for
various type parameters.

Closes https://github.com/roadster-rs/roadster/issues/546