feat!: Add `db-sea-orm` feature to prepare for other DB crate support (#612)
We'd like to add support for Diesel, so we should have a separate
feature flag for sea-orm vs diesel. Add a `db-sea-orm` flag and replace
all instances of `db-sql` with it. We'll keep the old `db-sql` feature
as well for now in case there's any common code to use between both, for
example, app config fields.