feat!: Add `diesel` support (#626)
This PR adds support for the
[Diesel](https://github.com/diesel-rs/diesel) SQL ORM. Diesel is a very
different ORM compared to SeaORM (the ORM we currently support), and as
such this PR required a lot of refactoring in order to provide a
relatively consistent experience regardless of which ORM a consumer
decides to use (or if they decide to use both, which is possible but not
particularly recommended). The refactor also resulted in some general
simplifications and improvements to the devx; read on for more details.

This PR contains a lot of breaking changes, the largest of which
include:

- Remove the `M` associated type from the `App` trait. A `Migrator` can
now be provided via the `migrators` method instead.
- Similarly, remove the `M` type parameter from `RoadsterApp`. SeaORM,
Diesel, or a generic `Migrator` can now be provided via the builder
methods.
- Change `RunCommand#run` to take a single `PreparedApp` struct
- ^This allowed removing the CLI handler method from the `AppService`
trait. CLI's now have access to the `ServiceRegistry` from the
`PreparedApp`, so they can get access to a particular `AppService` using
`ServiceRegistry#get` (assuming it was registered). Closes
https://github.com/roadster-rs/roadster/issues/617
- Consolidate DB migration CLI commands to provide consistent experience
between SeaORM and Diesel. This also removed some slightly redundant
commands.
- Rename  `AppContext#db` to `AppContext#sea_orm`
- Rename `App#db_connection_options` to
`App#sea_orm_connection_options`, and rename the related methods in
`RoadsterApp`
- Move/rename the DB health check
- Add `Sized` as a parent trait for the `App`

This PR also includes the following non-breaking changes:
- Add `AppContext` methods to get various Diesel connection pools types,
including Postgres, Mysql, Sqlite, and async pools for Postgres and
Mysql. Due to Diesel's type strategy for connections, there isn't a
single "DbConnection" like there is in SeaORM, so we provide individual
methods depending on which feature flags are enabled.
- Allow providing `AsyncSource` implementations to use with the `config`
crate. This allows, for example, loading secret config values from an
external service, such as AWS or GCS secrets managers. Closes
https://github.com/roadster-rs/roadster/issues/328
- Add a couple db config fields, `test-on-checkout` and
`retry-connection`
- Add more variants to our custom `Error` type

Closes https://github.com/roadster-rs/roadster/issues/625