feat: Create temporary databases for tests
Add an alternative to TestContainers for creating temporary databases:
create a new database using the original connection URI.

This approach may be preferable compared to TestContainers due to the
startup time (1-5 seconds usually) for each new TestContainer instance.
While using TestContainers does allow tests to run in parallel, this
start up time can cumulatively add up to be quite expensive in a large
test suite.

Instead of waiting for a new TestContainer instance to start for each
test, it's now possible to use the existing connection (e.g., a
persistent local docker instance) to create a new database for each test
which will be used for the duration of the test. The database can
optionally be deleted automatically when the test completes, if it
doesn't panic (enabled by default).

Closes https://github.com/roadster-rs/roadster/issues/634