fix!: Fix `run_test*` to skip CLI
The `run_test*` functions were parsing the CLI, which results in the
tests failing due to unexpected parameters.

Refactor some methods in `roadster::app` to allow skipping the CLI even
if the feature flag is enabled. This (for now) requires making the cli
fields optional in the prepared app. Maybe we can figure out a better
solution in the future.