chore!: Refactor app runners and change command + lifecycle params
Refactor the app runners (e.g., prepare/run/run_test/etc) into separate
modules. This cleans up the `roadster::app` module to just contain the
`App` trait definition and re-exports of the app runner functions.

This PR also changes the parameter type of the `RunCommand` trait
methods to a CLI-specific struct, and the `AppLifecycleHandler` trait
methods to a version of `PreparedApp` without the CLI structs.

Closes https://github.com/roadster-rs/roadster/issues/646