feat!: Allow customizing the Diesel pool connections
Add methods to the `App` trait and the `RoadsterApp` builder to allow
providing `[r2d2|bb8]::CustomizeConnection` implementations for each
supported connection pool type. Default implementations of the `App`
trait methods are provided, so this alone is not a breaking change.
However, this PR also makes the following breaking changes:

1. Creates a new `roadster::db` module and moves the
   `roadster::migration` module to be a submodule
2. Moves the diesel connection type aliases from
   `roadster::app::context` to `roadster::db`
(`roadster::db::migration`), which is a breaking change.