feat!: Pad snapshot case number with leading zeros in file name
If an rstest test switches from having <10 cases to having >=10 cases,
the case name switches to having a leading zero for cases 1-9, which
causes the snapshot file to change. To avoid this, update the snapshot
name to always be padded with a leading zero (for cases 1-9).

Also, write the testing and lifecycle hook chapters.