feat: Add `url.path` and `url.query` trace attributes (#738)
By default, all query params are redacted (but the param name is left
intact). Two new config fields allow configuring which params are
redacted:

- `service.http.middleware.tracing.query-params-allow-all`
- `service.http.middleware.tracing.query-param-names`

Also, fix setting the global default tracer when using `run_test` and
when running the tests via `cargo test` -- the `init_tracing` will not
fail the test if it's unable to set a the tracer due to it being set
already by a previous test run.