feat: Set HTTP request span name as `{method} {route}`
The OTEL semantic conventions specify that an HTTP request span should
be named `{method} {target}` if the method and target (route) if that
information is available. The `tracing-opentelemtry` crate allows
setting a dynamic span name using the `otel.name` field, which allows us
to use the method/route in the span name.

This PR also sets the http span kind as `SERVER`