feat!: Add `id` field to internal `Job` struct
To help with tracking background jobs in a uniform way across different
worker backends, add an `id` field to our internal `Job` struct. The ID
is logged at the debug level when the job is enqueued, at error level
when an error occurs, and is added to the trace span around the worker's
`handle` method.