feat!: Add `id` field to internal `Job` struct (#862)
To help with tracking background jobs in a uniform way across different
worker backends, add an `id` field to our internal `Job` struct. The ID
is logged at the debug level when the job is enqueued, at error level
when an error occurs, and is added to the trace span around the worker's
`handle` method.

This is technically a breaking change because any jobs that were
enqueued before this change will not deserialize properly due to the
missing `id` field. Additionally, our sidekiq integration was changed to
enqueue the entire `Job` struct instead of just the args.