fix!: Create separate `PeriodicJob` that's only needed for pg worker (#865)
We were biasing a little too much toward having a reusable `Job` struct
to use for all use cases. This made it a little awkward to use for the
periodic implementation in the pg worker, and added unnecessary data to
the jobs enqueued in the periodic implementation of the sidekiq worker.

This PR adds a new `PeriodicJob` struct that's very similar to the `Job`
struct, but adds a `periodic` field for the periodic data. This new
struct is only needed by the pg worker currently.

Because the periodic config is moved out of the `metadata` field, the
unique index and delete query for the periodic table needed to change.
This PR also removes an enum variant that is no longer needed. Both of
these require marking this as a breaking change.

Closes https://github.com/roadster-rs/roadster/issues/864