fix: Add default worker retry config values
Currently, workers will not retry by default even though the default
retry config has `max-retries = 25`. This is because there is no retry
strategy nor delay provided.

This PR adds sensible defaults for the retry config. This is maybe
technically a breaking change, but it can also be considered a bug fix
because we want workers to retry by default with the given `max-retries
= 25` default.

This PR also fixes the exponential backoff delay calculation -- before,
it would always be 0 if the delay config was < 1000 and always 1 second
if the delay config was >= 1000 and < 2000.

This PR also fixes the delay calcuations to use `saturating_*` methods
to avoid overflows.