chore!: Rearrange method params to put state as first param
Some methods use the app state as the first parameter and others do not.
This PR refactors all methods to ensure that any method that takes app
state as a parameter has it as the first parameter. The exception is the
few methods that take an instance of the `App` trait -- those take the
`App` trait as the first parameter.

Closes https://github.com/roadster-rs/roadster/issues/934